requirements.txt
alembic==1.14.0
click==8.1.8
Flask==2.0.1
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==20.1.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
Mako==1.3.8
MarkupSafe==2.1.5
psycopg2-binary==2.9.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
typing-extensions==4.12.2
Werkzeug==2.0.3
zipp==3.20.2
Flask-Login==0.6.2
